+++ NEWS.md
@@ -9,6 +9,7 @@ border-radius: 128px;
 ## v1.9.23
 
 - Fixes
+    - Fix admin tasks tab crash
     - Fix Library folder picker showing bad data
 
 ## v1.9.22